refactor: replace inline empty states with EmptyState component
