+++ src/components/overview/recent-tasks/recent-tasks-table.tsx
@@ -4,7 +4,7 @@ import { useState } from 'react'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
 import Link from 'next/link'
 import { AppRoutes } from '@/constants/routes'
-import { TaskDialog } from '@/components/tasks/task-dialog'
+import { TaskDialog } from '@/components/tasks/task-dialog/task-dialog'
 import { RecentTasksCards } from './cards'
 import { RecentTasksTableView } from './table-view'
 import type { RecentTask } from '@/types/task'
